backend/requirements.txt
numpy==1.26.4
passlib== 1.7.4
aiofiles==23.2.1
alembic==1.16.1
anyio==4.9.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
async-timeout==5.0.1
asyncpg==0.29.0
bcrypt==4.3.0
certifi==2025.6.15
cffi==1.17.1
click==8.2.1
cryptography==45.0.4
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
exceptiongroup==1.3.0
fastapi==0.110.0
fastapi-users==14.0.1
fastapi-users-db-sqlalchemy==7.0.0
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.27.2
idna==3.10
makefun==1.16.0
Mako==1.3.10
MarkupSafe==3.0.2
openai==1.26.0
psycopg2-binary==2.9.10
pwdlib==0.2.1
pyasn1==0.6.1
pycparser==2.22
pydantic==1.10.22
PyJWT==2.10.1
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.20
redis==6.2.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.30
starlette==0.36.3
tomli==2.2.1
tqdm==4.67.1
typing_extensions==4.14.0
uvicorn==0.29.0
tiktoken==0.9.0
chromadb==0.4.14
celery==5.5.3
elevenlabs==2.7.1
ffmpeg-python==0.2.0
elevenlabs
celery[redis]
ffmpeg-python
